Fix multiple transaction ID assignment bug.
In Node.js we can fire off multiple parallel HTTP requests. What this means for Roach.js is that TxnSender may send multiple requests for the same transaction with Txn.ID = null. That causes the server to create multiple transactions, releasing Zalgo.

The fix is to queue requests at least until we have a Txn.ID to use.

There is still a lack of transactionality if we don't serialize requests per transaction. We'll aim to address this in a future patch.