Fix handling of remote errors in transactions.
The previous code threw plain objects, which is an anti-pattern. This patch creates a custom error class that we can use.

There were some bugs related to handling errors in the Prepare callback instead of in the Flush callback.

Updated the restart logic to obey the server-side recommendation (transaction_restart).